move codingplans to configurations
