+++ src/analysis_file.py
@@ -10,7 +10,7 @@ from core_data_modules.traced_data.util.fold_traced_data import FoldStrategies
 from core_data_modules.util import TimeUtils
 
 from src.lib import PipelineConfiguration, ConsentUtils, ListeningGroups
-from src.lib.pipeline_configuration import CodingModes
+from src.lib.configuration_objects import CodingModes
 
 
 class AnalysisFile(object):
@@ -76,7 +76,7 @@ class AnalysisFile(object):
         # dadaab does not have listening groups.
         if pipeline_configuration.pipeline_name == "kakuma_pipeline":
             export_keys.append("repeat_listening_group_participant")
-            for plan in PipelineConfiguration.KAKUMA_RQA_CODING_PLANS:
+            for plan in PipelineConfiguration.RQA_CODING_PLANS:
                 export_keys.extend([f'{plan.dataset_name}_listening_group_participant'])
         else:
             assert pipeline_configuration.pipeline_name == "dadaab_pipeline", "PipelineName must be either " \
